Use yggdrasil as source-of-truth for finding charts to mirror
The old way of doing this was mirroring the Helm chart git repository
defined in the manifest file.

The new way has a few advantages:
- Support for charts in different chart repositories
- Only required charts are mirrored and templated (much faster)
- Support for skipping broken charts can be removed, as all charts
  defined in yggdrasil should be deployable

This also changes the "directory structure" for charts, as all the
charts are mirrored into a single Helm chart repository (was a bit
easier to implement). This requires a change to fwd-proxy, which can be
found here[1].

[1] https://github.com/distributed-technologies/helm-charts/pull/247