Support incremental mirroring for images
This adds a `-i` (`--incremental`) option to the `mirror-images`
command, which only mirror images without `skip: true` in the manifest
file and set `skip: true` for the images it mirrors. It is up to the
user to keep track of which images have been deployed.

The use case is to save bandwidth, as transferring the tarball with all
the images over the network every time is wasteful and slow.