Support specifying a branch instead of a commit in config.json
This way config.json doesn't need to be updated everytime a commit is
pushed to ex: main. The commit id is still noted in the manifest file.

This also allows a workflow where syncctl is run from CI on every push
(to ex: main) and perhaps commits the manifest file automatically to
another branch.