Support overriding the Git URL used to detect references to the cloned repo
In the commit[1] switching to Flux, the following is stated:
"Git repositories (GitRepository[3]) specified in Flux aren't packaged,
as you are expected to run syncctl on every individuel repository (ex:
to support different release cadence in a multi tenancy[2] setup)."

This means syncctl isn't cloning referred GitRepository[3], but syncctl
must still be able to detect a GitRepository[3] referring to the current
repository, so it knows which Flux Kustomizations to template.

Some Git providers unfortunately provides different Git URLs for HTTP
and SSH cloning (ex: Azure DevOps), this commit adds a option for
overriding the Git URL to compare against (syncctl is already ignoring
protocol, domain and basic auth credentials when comparing URLs).

[1] 96a28d1 ("Support Flux instead of Yggdrasil[1]/Argo CD")
[2] https://github.com/fluxcd/flux2-multi-tenancy
[3] https://fluxcd.io/docs/components/source/gitrepositories/