Avoid hard coding the git repos by creating them on-demand
In [1] the git repositories was changed to be namespaced, but the git
server doesn't support creating a new git repository on push, and we
don't want to hardcode the list of git repositories. So tweak the
sync.sh script to create the git repository on disk, if it doesn't
exist. Also fix wrong "git path" which should have been adjusted in[1].

[1] 96a28d1 ("Support Flux instead of Yggdrasil[1]/Argo CD")

Fixes: 96a28d1 ("Support Flux instead of Yggdrasil[1]/Argo CD")